build: Updating mediawiki/mediawiki-phan-config to 0.9.0
Depends-On: Ief9a9d19bbeecd29ab13244e18eeabc0aa08a520
Depends-On: Ic972fcafd142784277e131b25f08d9dc2cb84781

+++ includes/ProcessBounceEmails.php
@@ -116,7 +116,7 @@ abstract class ProcessBounceEmails {
 
 		$failedUser = [];
 
-		$currentTime = wfTimestamp();
+		$currentTime = (int)wfTimestamp();
 		preg_match( '~(.*?)@~', $hashedEmail, $hashedPart );
 		if ( !isset( $hashedPart[1] ) ) {
 			wfDebugLog( 'BounceHandler',
@@ -136,7 +136,7 @@ abstract class ProcessBounceEmails {
 			);
 			return [];
 		}
-		$bounceTime = base_convert( $hashedVERPPart[3], 36, 10 );
+		$bounceTime = (int)base_convert( $hashedVERPPart[3], 36, 10 );
 		// Check if the VERP hash is valid
 		if ( base64_encode(
 				substr( hash_hmac( $wgVERPalgorithm, $hashedData, $wgVERPsecret, true ), 0, 12 )
